+++ app/src/main/java/com/acg/outtamycircle/entitycomponent/impl/gameobjects/GameCharacter.java
@@ -1,5 +1,5 @@
 package com.acg.outtamycircle.entitycomponent.impl.gameobjects;
 
 public class GameCharacter extends GameObject{
-    public GameCharacter(short id){ super(id, Type.CHARACTER); }
+    public GameCharacter(int id){ super(id, Type.CHARACTER); }
 }
