+++ app/src/main/java/com/acg/outtamycircle/entitycomponent/impl/gameobjects/GameObject.java
@@ -3,17 +3,17 @@ package com.acg.outtamycircle.entitycomponent.impl.gameobjects;
 import com.acg.outtamycircle.entitycomponent.Entity;
 
 public class GameObject extends Entity {
-    private final short objectId;
+    private final int objectId;
     private final Type type;
 
     public enum Type {CHARACTER, POWERUP}
 
-    public GameObject(short objectId, Type type){
+    public GameObject(int objectId, Type type){
         this.objectId = objectId;
         this.type = type;
     }
 
-    public short getObjectId(){ return objectId; }
+    public int getObjectId(){ return objectId; }
 
     public Type getType() {
         return type;
