+++ app/src/main/java/com/acg/outtamycircle/utilities/IdGenerator.java
@@ -3,18 +3,18 @@ package com.acg.outtamycircle.utilities;
 public class IdGenerator {
     private static IdGenerator instance = null;
 
-    private IdGenerator(short id) {
+    private IdGenerator(int id) {
         this.id = id;
     }
 
-    public static IdGenerator getInstance(short startId){
+    public static IdGenerator getInstance(int startId){
         if(instance == null)
             instance = new IdGenerator(startId);
         return instance;
     }
 
-    private short id;
-    public short next(){
+    private int id;
+    public int next(){
         return ++id;
     }
 }
